requirements.txt
# Core dependencies
requests>=2.31.0
pydub>=0.25.1
mutagen>=1.47.0
python-dotenv>=1.0.0

# Cloud storage
azure-storage-blob>=12.19.0

# GUI framework
customtkinter>=5.2.0
pillow>=10.0.0

# Audio processing
yt-dlp>=2023.7.6
imageio-ffmpeg>=0.4.8

# Web scraping and API
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Steam API (if available)
steam>=1.4.4

# File handling
pathlib2>=2.3.7

# Configuration
configparser>=6.0.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
